# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: RelayFlow API
  version: 0.1.0
  description: API contract for the RelayFlow backend.
servers:
  - url: http://localhost:8080
paths:
...
  /workspaces:
    get:
      operationId: listWorkspaces
      summary: List workspaces the authenticated user is a member of
      tags: [Workspaces]
      responses:
        "200":
          description: Workspace list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkspaceResponse"
    post:
      operationId: createWorkspace
      summary: Create a new workspace
      tags: [Workspaces]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkspaceRequest"
      responses:
        "201":
          description: Workspace created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceResponse"
  /workspaces/{workspaceId}:
    patch:
      operationId: updateWorkspace
      summary: Rename a workspace
      description: Owner-only endpoint.
      tags: [Workspaces]
      parameters:
        - $ref: "#/components/parameters/workspaceIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWorkspaceRequest"
      responses:
        "200":
          description: Workspace updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceResponse"
  /workspaces/{workspaceId}/contact-field-definitions:
    put:
      operationId: updateContactFieldDefinitions
      summary: Define the workspace's custom contact field schema
      description: >-
        Requires the `CONTACT_FIELDS_WRITE` permission or owner role. A field's key can't collide
        with a reserved key (`displayName`, `phone`, `email`, `country`) — RelayFlow already fills
        those in automatically and they can't be redefined as a custom field.
      tags: [Workspaces]
      parameters:
        - $ref: "#/components/parameters/workspaceIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateContactFieldDefinitionsRequest"
      responses:
        "200":
          description: Workspace updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceResponse"
        "400":
          description: A field key collides with a reserved key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /workspaces/{workspaceId}/members:
    get:
      operationId: listWorkspaceMembers
      summary: List members of a workspace
      tags: [WorkspaceMembers]
      parameters:
        - $ref: "#/components/parameters/workspaceIdPath"
      responses:
        "200":
          description: Workspace member list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkspaceMemberResponse"
    post:
      operationId: inviteWorkspaceMemberDirectly
      summary: Add or invite a workspace member directly by email
      tags: [WorkspaceMembers]
      parameters:
        - $ref: "#/components/parameters/workspaceIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InviteMemberRequest"
      responses:
        "201":
          description: Workspace member added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceMemberResponse"
        "409":
          description: User is already a member of this workspace
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /workspaces/{workspaceId}/members/{memberId}:
    patch:
      operationId: updateWorkspaceMember
      summary: Update a workspace member role or permissions
      description: >
        Owner-only endpoint. Direct promotion to OWNER is rejected; ownership must be
        transferred through `/workspaces/{workspaceId}/owner`.
      tags: [WorkspaceMembers]
      parameters:
        - $ref: "#/components/parameters/workspaceIdPath"
        - name: memberId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMemberRequest"
      responses:
        "200":
          description: Workspace member updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceMemberResponse"
        "409":
          description: Invalid role update, such as direct owner promotion or owner demotion
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: removeWorkspaceMember
      summary: Remove a workspace member
      tags: [WorkspaceMembers]
      parameters:
        - $ref: "#/components/parameters/workspaceIdPath"
        - name: memberId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Workspace member removed
        "409":
          description: Cannot remove the workspace owner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /workspaces/{workspaceId}/owner:
    put:
      operationId: transferWorkspaceOwnership
      summary: Transfer workspace ownership to another member
      description: >
        Owner-only endpoint. The current owner is demoted to MEMBER and the target
        member becomes OWNER. The target must already be a member of the workspace.
      tags: [WorkspaceMembers]
      parameters:
        - $ref: "#/components/parameters/workspaceIdPath"
        - name: memberId
          in: query
          required: true
          schema:
            type: string
            format: uuid
          description: Workspace member ID that should become the new owner.
      responses:
        "204":
          description: Ownership transferred
        "409":
          description: Invalid transfer, such as transferring ownership to yourself
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Only the current workspace owner can transfer ownership
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Target member not found in the workspace
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /channel-accounts:
    get:
      operationId: listChannelAccounts
      summary: List channel accounts for a workspace
      tags: [ChannelAccounts]
      parameters:
        - $ref: "#/components/parameters/workspaceIdQuery"
      responses:
        "200":
          description: Channel account list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChannelAccountResponse"
    post:
      operationId: createChannelAccount
      summary: Connect a new channel account
      tags: [ChannelAccounts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateChannelAccountRequest"
      responses:
        "201":
          description: Channel account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChannelAccountResponse"
        "403":
          description: Missing CHANNELS_WRITE permission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /channel-accounts/{id}:
    delete:
      operationId: disconnectChannelAccount
      summary: Disconnect a channel account (sets status to DISABLED; data is preserved)
      tags: [ChannelAccounts]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: "#/components/parameters/workspaceIdQuery"
      responses:
        "204":
          description: Channel account disconnected
        "404":
          description: Channel account not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /channel-accounts/{id}/reconnect:
    post:
      operationId: reconnectChannelAccount
      summary: Re-enable a disconnected channel account (sets status back to ACTIVE)
      tags: [ChannelAccounts]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: "#/components/parameters/workspaceIdQuery"
      responses:
        "200":
          description: Channel account re-enabled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChannelAccountResponse"
        "404":
          description: Channel account not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Contacts ────────────────────────────────────────────────────────────────
  /contacts:
    get:
      operationId: listContacts
      summary: List contacts for a workspace, newest first
      tags: [Contacts]
      parameters:
        - $ref: "#/components/parameters/workspaceIdQuery"
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Paginated contact list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactPageResponse"
    post:
      operationId: createContact
      summary: Create a contact in a workspace
      tags: [Contacts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContactRequest"
      responses:
        "201":
          description: Contact created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactResponse"
  /contacts/{id}:
    get:
      operationId: getContact
      summary: Get a contact with linked channel identities
      description: >-
        The response's `customFields` merges explicitly-stored values with an auto-derived
        value for any reserved key (displayName, phone, email, country) that hasn't been
        explicitly set — e.g. `phone` from a WhatsApp identity. An explicitly-stored value
        always wins over the derived one.
      tags: [Contacts]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: "#/components/parameters/workspaceIdQuery"
      responses:
        "200":
          description: Contact detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactDetailResponse"
        "404":
          description: Contact not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteContact
      summary: Delete a contact
      tags: [Contacts]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: "#/components/parameters/workspaceIdQuery"
      responses:
        "204":
          description: Contact deleted
        "404":
          description: Contact not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /contacts/{id}/custom-fields:
    patch:
      operationId: updateContactCustomFields
      summary: Update a contact's custom field values
      description: >-
        Requires the `CONTACT_FIELDS_WRITE` permission or owner role. Accepts the full set of
        field values (reserved and workspace-defined); the response's `customFields` merges these
        explicitly-stored values with any auto-derived reserved values (e.g. phone from a
        WhatsApp identity) that weren't explicitly overridden.
      tags: [Contacts]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: "#/components/parameters/workspaceIdQuery"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateContactCustomFieldsRequest"
      responses:
        "200":
          description: Contact detail with updated custom fields
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactDetailResponse"
        "404":
          description: Contact not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /contacts/{id}/merge:
    post:
      operationId: mergeContacts
      summary: Merge a source contact into a target contact
      tags: [Contacts]
      parameters:
        - name: id
          in: path
          required: true
          description: Target contact ID that will remain after the merge
          schema:
            type: string
            format: uuid
        - $ref: "#/components/parameters/workspaceIdQuery"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MergeContactRequest"
      responses:
        "200":
          description: Updated target contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactResponse"
        "409":
          description: Invalid merge request, such as merging a contact with itself or merging contacts with identities on the same channel account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Contact not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /external-identities:
    post:
      operationId: createExternalIdentity
      summary: Link a contact to an external channel identity
      tags: [Contacts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExternalIdentityRequest"
      responses:
        "201":
          description: External identity created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExternalIdentityResponse"

  # ── Conversations ─────────────────────────────────────────────────────────────
  /conversations:
    get:
      operationId: listConversations
      summary: List conversations for a workspace, newest first
      tags: [Conversations]
      parameters:
        - $ref: "#/components/parameters/workspaceIdQuery"
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 30
        - name: contactId
          in: query
          required: false
          description: When present, only conversations for this contact are returned
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Paginated conversation list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversationPageResponse"
    post:
      operationId: createConversation
      summary: Create a conversation
      tags: [Conversations]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateConversationRequest"
      responses:
        "201":
          description: Conversation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversationResponse"
  /conversations/{conversationId}:
    get:
      operationId: getConversation
      summary: Get a single conversation
      tags: [Conversations]
      parameters:
        - $ref: "#/components/parameters/conversationIdPath"
        - $ref: "#/components/parameters/workspaceIdQuery"
      responses:
        "200":
          description: Conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversationResponse"
        "404":
          description: Conversation not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateConversation
      summary: Update a conversation status
      tags: [Conversations]
      parameters:
        - $ref: "#/components/parameters/conversationIdPath"
        - $ref: "#/components/parameters/workspaceIdQuery"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateConversationRequest"
      responses:
        "200":
          description: Conversation updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversationResponse"

  # ── Messages ──────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
  parameters:
    workspaceIdPath:
      name: workspaceId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    CreateWorkspaceRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
    UpdateWorkspaceRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
    UpdateContactFieldDefinitionsRequest:
      type: object
      required: [contactFieldDefinitions]
      properties:
        contactFieldDefinitions:
          type: array
          items:
            $ref: "#/components/schemas/ContactFieldDefinition"
...
    ErrorResponse:
      type: object
      required: [message, timestamp]
      properties:
        message:
          type: string
        timestamp:
          type: string
          format: date-time
...
    WorkspaceResponse:
      type: object
      required: [id, name, contactFieldDefinitions, createdAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        contactFieldDefinitions:
          type: array
          items:
            $ref: "#/components/schemas/ContactFieldDefinition"
        createdAt:
          type: string
          format: date-time
...
    WorkspaceMemberResponse:
      type: object
      required: [id, userId, role, permissions, joinedAt]
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        email:
          type: string
          format: email
          nullable: true
        displayName:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        role:
          $ref: "#/components/schemas/WorkspaceRole"
        permissions:
          type: array
          uniqueItems: true
          items:
            $ref: "#/components/schemas/WorkspacePermission"
        joinedAt:
          type: string
          format: date-time
...